feat: SSR 제거
