+++ src/app/(posts)/[slug]/page.tsx
@@ -8,8 +8,6 @@ import Header from '@/components/post/header';
 import Reaction from '@/components/reaction';
 import config from '@/configs/config.json';
 import { accessPost, getPost, getPosts } from '@/lib/MDX';
-import { getCommentBySlug } from '@/lib/supabase/comment.server';
-import { getReactionBySlug } from '@/lib/supabase/reaction.server';
 
 interface PageProps {
   params: { slug: string };
@@ -21,15 +19,12 @@ const Page = async ({ params }: PageProps) => {
   const { frontmatter, toc, MDX } = await getPost(params.slug);
   const { title, date, tags } = frontmatter;
 
-  const reactions = await getReactionBySlug(params.slug);
-  const comments = await getCommentBySlug(params.slug);
-
   return (
     <Actions>
       <Header title={title} date={date} tags={tags} />
       <Post toc={toc} MDX={MDX} />
-      <Reaction data={reactions} slug={params.slug} />
-      <Comment data={comments} slug={params.slug} />
+      <Reaction slug={params.slug} />
+      <Comment slug={params.slug} />
     </Actions>
   );
 };
