+++ src/components/comment/index.tsx
@@ -1,23 +1,21 @@
 'use client';
 
-import { useCallback, useState } from 'react';
+import { useCallback, useEffect, useState } from 'react';
 
 import Read from '@/components/comment/read';
 import Write from '@/components/comment/write';
 import useActions from '@/hooks/use-actions';
-import { postComment } from '@/lib/supabase/comment.client';
+import { getCommentBySlug, postComment } from '@/lib/supabase/comment.client';
+import type { Comment } from '@/lib/supabase/comment.client';
 import { getItem } from '@/utils/local-storage';
 import { burst } from '@/utils/particle';
 
 interface Props {
-  data: Comment[];
   slug: string;
 }
 
-export type Comment = { id: string; emoji: string; label: string; value: string };
-
-const Comment = ({ data, slug }: Props) => {
-  const [comments, setComments] = useState(data);
+const Comment = ({ slug }: Props) => {
+  const [comments, setComments] = useState<Comment[]>([]);
   const { hasActions, setActions } = useActions();
 
   const handleWrite = useCallback(
@@ -40,6 +38,10 @@ const Comment = ({ data, slug }: Props) => {
     },
     [hasActions, setActions, slug]
   );
+
+  useEffect(() => {
+    getCommentBySlug(slug).then((data) => setComments(data));
+  }, [slug]);
 
   return (
     <section className="space-y-4">
