+++ src/lib/supabase/comment.client.ts
@@ -1,5 +1,19 @@
 import { supabase } from '@/lib/supabase/client-csr';
 
+export type Comment = { id: string; emoji: string; label: string; value: string };
+
+export const getCommentBySlug = async (slug: string) => {
+  const { data, error } = await supabase
+    .from('comments')
+    .select('id,emoji,label,value')
+    .eq('slug', slug)
+    .order('created_at', { ascending: false });
+
+  if (error) throw error;
+
+  return data;
+};
+
 export const postComment = async (
   user_id: string,
   slug: string,
